Move CIABEligibilityChecker to Yosemite
