+++ Modules/Sources/Yosemite/Tools/CIAB/CIABEligibilityCheckerProtocol.swift
@@ -1,8 +1,6 @@
 import Foundation
-import Yosemite
 
-/// periphery: ignore - Will be used in upcoming changes for app feature gating
-protocol CIABEligibilityCheckerProtocol {
+public protocol CIABEligibilityCheckerProtocol {
     var isCurrentSiteCIAB: Bool { get }
 
     func isSiteCIAB(_ site: Site) -> Bool
